requirements.txt
aiohttp==3.7.4
appdirs==1.4.4
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
black==20.8b1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
coverage==4.5.1
coveralls==1.3.0
discord.py==1.6.0
docopt==0.6.2
idna==2.10
importlib-metadata==3.7.0
more-itertools==8.7.0
multidict==5.1.0
mypy-extensions==0.4.3
parsedatetime==2.6
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pytest==3.7.1
pytest-cov==2.5.1
python-dotenv==0.15.0
pytz==2021.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
timefhuman==0.0.5
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
zipp==3.4.1
